backend/livedata_worker/requirements.txt
influxdb-client==1.39.0
python-dotenv==1.0.0
websockets==12.0